refactor: share stubCustomerRepo across tests
